+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Fixed
+ - Set `WINEDLLPATH` and `WINELOADER` environment variables.
+
 ## [1.2.1] - 2019-04-08
 ### Changed
  - Delay Proton detection until it's necessary.